[#2] fix : Atoms Component type 수정
