--- types/family.tsx
@@ -1,5 +1,0 @@
-export type FamilyCardProps = {
-  familyName: string;
-  familyMember: string[];
-  color: string;
-};
