--- types/register.tsx
@@ -1,3 +1,0 @@
-export type RegisterCardProps = {
-    text: string;
-  };
