--- types/role.tsx
@@ -1,4 +1,0 @@
-export type RoleCardProps = {
-  image: string;
-  name: string;
-};
